requirements.txt
typer
pyyaml
telethon
pywebcopy
